# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type CaseFileStatus = typeof CaseFileStatus[keyof typeof CaseFileStatus];


export const CaseFileStatus = {
  active: 'active',
  review: 'review',
  completed: 'completed',
} as const;

export interface CaseFile {
  id: number;
  title: string;
  /** @nullable */
  clientName: string | null;
  matterType: string;
  jurisdiction: string;
  status: CaseFileStatus;
  progress: number;
  /** @nullable */
  summary: string | null;
  issues: string[];
  keyFacts: string[];
  nextSteps: string[];
  /** @nullable */
  nextDeadline: string | null;
  messagesCount: number;
  createdAt: string;
  updatedAt: string;
}

export interface CaseFileInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  matterType: string;
  /** @nullable */
  clientName?: string | null;
  jurisdiction: string;
}

export type CaseFileUpdateStatus = typeof CaseFileUpdateStatus[keyof typeof CaseFileUpdateStatus];


export const CaseFileUpdateStatus = {
  active: 'active',
  review: 'review',
  completed: 'completed',
} as const;

export interface CaseFileUpdate {
  /** @minLength 1 */
  title?: string;
  /** @nullable */
  clientName?: string | null;
  matterType?: string;
  jurisdiction?: string;
  status?: CaseFileUpdateStatus;
  progress?: number;
  /** @nullable */
  summary?: string | null;
  issues?: string[];
  keyFacts?: string[];
  nextSteps?: string[];
  /** @nullable */
  nextDeadline?: string | null;
}

export type CaseFileMessageRole = typeof CaseFileMessageRole[keyof typeof CaseFileMessageRole];


export const CaseFileMessageRole = {
  user: 'user',
  assistant: 'assistant',
} as const;

export interface CaseFileMessage {
  id: number;
  caseFileId: number;
  role: CaseFileMessageRole;
  content: string;
  createdAt: string;
}

export interface CaseFileMessageInput {
  /** @minLength 1 */
  content: string;
}

export interface CaseFileActivity {
  id: number;
  title: string;
  activity: string;
  updatedAt: string;
}

export interface CaseFilesSummary {
  total: number;
  active: number;
  review: number;
  completed: number;
  recentActivity: CaseFileActivity[];
}