# lib/api-zod/src/generated/types/caseFile.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CaseFileStatus } from './caseFileStatus';

export interface CaseFile {
  id: number;
  title: string;
  /** @nullable */
  clientName: string | null;
  matterType: string;
  jurisdiction: string;
  status: CaseFileStatus;
  progress: number;
  /** @nullable */
  summary: string | null;
  issues: string[];
  keyFacts: string[];
  nextSteps: string[];
  /** @nullable */
  nextDeadline: Date | null;
  messagesCount: number;
  createdAt: Date;
  updatedAt: Date;
}