# lib/api-zod/src/generated/types/caseFileActivity.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CaseFileActivity {
  id: number;
  title: string;
  activity: string;
  updatedAt: Date;
}