# lib/api-zod/src/generated/types/caseFileInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CaseFileInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  matterType: string;
  /** @nullable */
  clientName?: string | null;
  jurisdiction: string;
}