# lib/api-zod/src/generated/types/caseFileMessage.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CaseFileMessageRole } from './caseFileMessageRole';

export interface CaseFileMessage {
  id: number;
  caseFileId: number;
  role: CaseFileMessageRole;
  content: string;
  createdAt: Date;
}