# lib/api-zod/src/generated/types/caseFileMessageInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CaseFileMessageInput {
  /** @minLength 1 */
  content: string;
}