# lib/api-zod/src/generated/types/caseFileUpdate.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CaseFileUpdateStatus } from './caseFileUpdateStatus';

export interface CaseFileUpdate {
  /** @minLength 1 */
  title?: string;
  /** @nullable */
  clientName?: string | null;
  matterType?: string;
  jurisdiction?: string;
  status?: CaseFileUpdateStatus;
  progress?: number;
  /** @nullable */
  summary?: string | null;
  issues?: string[];
  keyFacts?: string[];
  nextSteps?: string[];
  /** @nullable */
  nextDeadline?: Date | null;
}