# lib/api-zod/src/generated/types/caseFilesSummary.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CaseFileActivity } from './caseFileActivity';

export interface CaseFilesSummary {
  total: number;
  active: number;
  review: number;
  completed: number;
  recentActivity: CaseFileActivity[];
}